Merge changes from presto-admin Dockerfile into centos6-ssh-oj8
The genesis of the docker cluster was to be able to run the presto-admin
product tests for yarn-slider on docker. I copied the docker file from
presto-admin so as to be able develop the docker cluster without interfering
with the bpresto-admin product tests that don't require hadoop (i.e., most of
them).

It's finally time to kill off the Dockerfile in presto-admin and bring the
product tests onto the centos6-ssh-oj8 image. Before doing so, we need to merge
over a couple of improvements from the Dockerfile:

0) Fail the container build if the insecure private key is missing. This has
been made redundant by the next item, but we should still fail fast if we screw
things up somehow in the future.

1) Put the insecure private key in the repo rather than fetching it from the
internet. Network issues (ours and theirs) were causing spurious intermittent
failures.

2) Reduce the number of layers we create installing things with yum. I
consolidated a few more commands in this commit to further reduce the number of
useless layers we create.