Fix namenode not starting error in hdp2.5-hive
Previously the namenode would be restarted by
supervisor.d because of exiting due to the error

org.apache.hadoop.hdfs.server.common.InconsistentFSStateException: Directory $DIR is in an inconsistent state: storage directory does not exist or is not accessible.

where $DIR was something along the lines of

/tmp/hadoop/dfs/name

This commit fixes it as suggested in
http://stackoverflow.com/questions/6447885/no-namenode-error-in-pseudo-mode

Most likely, the default namenode dir has changed between verisons.