Update jdeathe/centos-ssh base image
The centos-6-1.2.0 tag was causing problems when we went to install tar
during the docker build. The problem seems to be related to the tar rpm
from the yum repo, so we updated the base image.

Tar now installs correctly, but further changes were needed to
accomodate the changes to the sshd bootstrap scripts.