Remove dependency on docker-release
In the last 10 months, neither the Makefile nor the docker-release code
have changed. Unfortunately, something in the world has and we now have
2 problems:

1. docker-release has a dependency on docker-py which causes problems
for docker-compose.
2. docker-release fails to push to dockerhub with some kind of
permissions error.

This change implements the functionality formerly in docker-release(*)
using make, git, shell, and native docker commands. This frees us from
depending on the state of the world in python when actually releasing
images.

(*) This does not implement checking for an existing tag on dockerhub
that matches the one we're trying to push. Experience has shown that
that check frequently results in having to finish a push manually when
docker-release's push gets interrupted. In that case, retrying via
docker-release fails because some image with the tag has already been
pushed.