Update Zulu JDK to version 17 for CDH5.15
We want to start testing JDK 17 compatibility in the Trino project
without a need to move all images to JDK 17 at once. CDH 5.15 was
selected as it has both kerberized and KMS environments, which will
allow us to test i.e. Kerberos implementation. For the time being,
we will have product tests running on both JDK 11 and JDK 17 at
the same time.